:art: art : useSearchParams에 따른 컴포넌트 분리
